aiodownloader/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
instaloader==4.9.6
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.4.0
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
instaloader==4.9.6
lxml==4.9.2
MechanicalSoup==1.2.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.4.0
